Firebase day2

+++ app/src/main/java/com/lavish/android/practice/models/Product.java
@@ -10,7 +10,7 @@ public class Product implements Serializable {
 
     //Compulsory
     public String name;
-    public byte type;
+    public int type;
 
     //WeightBased
     public int pricePerKg;
